requirements.txt
datasets
pyarrow
Pillow
requests
pytesseract
chromadb
duckdb
duckdb-engine
tqdm
transformers
torch
accelerate
torchvision
sglang
pybase64
pydantic
qwen_vl_utils
zmq
uvloop
fastapi
sentencepiece
compressed_tensors
openai
partial_json_parser
einops
sgl_kernel
sentence_transformers
nltk
rouge-score
bert-score
ipywidgets
nest_asyncio
bitsandbytes